[OT-233][FIX]: 대시보드 페이지 타입 위치 변경
