+++ src/entities/dashboard/api/tagRecommendPlaylistApi.ts
@@ -1,5 +1,22 @@
 import { api } from "@shared/api";
-import { ApiResponse, TagPlaylistResponse } from "@shared/types";
+import { ApiResponse, PageInfo } from "@shared/types";
+
+// 해당 태그 모달창 아래 뜨는 콘텐츠 안쪽 타입
+export interface TagPlaylistItem {
+  mediaId: number;
+  title: string;
+  posterUrl: string;
+  thumbnailUrl: string;
+  mediaType: "CONTENTS" | "SERIES";
+  duration: number;
+  positionSec: number;
+}
+
+// 해당 태그 모달창 추천 콘텐츠 리스트 전체 타입
+export interface TagPlaylistResponse {
+  pageInfo: PageInfo;
+  dataList: TagPlaylistItem[];
+}
 
 export const tagRecommendPlaylistApi = {
   getTagRecommendPlaylist: async (tagId: number) =>
